Polishing Demo #3
- Switched Base Damage value for Characters to a float (from int8) to
simplify math
- Added Scale Interface to control player/item behavior based off of
player/item size ratio (Sets Total Damage and Throw Speed based off of
ratio)
- Refined Blunt Damage logic for AI, causing damage to the multiplied
based off of scale difference (needs to be updated further)
- Doors have collision
- Impale Should do damage again

+++ Source/BrensFullGame/BCharacter.cpp
@@ -145,7 +145,7 @@ void ABCharacter::SetValues(AItem* Item, uint8 TypeOfItem, bool ItemHeld)
 	IsItemHeld = ItemHeld;
 }
 
-void ABCharacter::SetPlayerDamage(uint8 BaseDamage, uint8 DamageType, float Duration, AItem* Hitter)
+void ABCharacter::SetPlayerDamage(float BaseDamage, uint8 DamageType, float Duration, AItem* Hitter)
 {
 	DamageBase = BaseDamage;
 	AgentType = DamageType;
